update build workflow

+++ housekeeping/build.sh
@@ -9,19 +9,39 @@ RUSTFLAGS='-Dwarnings'
 RUNTIME_DIR='runtime'
 allfeatures='private-net,wip,ready-to-test'
 
-# build func
-releasetest(){
-    cargo test --release --features "private-net runtime-benchmarks"
-}
+test
+if [[ $buildTag != null ]] && [[ ${TAG_NAME} != null || ${TAG_NAME} != '' ]]; then
+    if [[ ${TAG_NAME} =~ 'benchmarking'* ]]; then
+        build 'private-net runtime-benchmarks' 0
+    elif [[ ${TAG_NAME} =~ 'stage'* ]]; then
+        build 'private-net include-real-files ready-to-test' 0
+    elif [[ ${TAG_NAME} =~ 'test'* ]]; then
+        build 'private-net include-real-files reduced-pswap-reward-periods ready-to-test' 0
+    elif [[ -n ${TAG_NAME} && ${TAG_NAME} != 'predev' ]]; then
+        build 'include-real-files' 101
+    fi
+else
+    if [ $prBranch = 'master' ]; then
+        printf "⚡️ Running tests and migrations %s\n"
+        RUST_LOG="debug cargo test --features try-runtime -- run_migrations"
+    fi
+fi
 
-test(){
-    printf "⚡️ Running Tests for code coverage only %s\n"
-    export RUSTFLAGS="-Cinstrument-coverage"
-    export SKIP_WASM_BUILD=1
-    export LLVM_PROFILE_FILE="sora2-%p-%m.profraw"
-    rm -rf ~/.cargo/.package-cache
-    cargo fmt -- --check > /dev/null
-    cargo test --features $allfeatures
+# build func
+test() {
+    if [[ $buildTag != null ]] && [[ ${TAG_NAME} != null || ${TAG_NAME} != '' ]]; then
+        printf "⚡️ Tag is %s\n" $buildTag ${TAG_NAME}
+        printf "⚡️ Testing with features: private-net runtime-benchmarks\n"
+        cargo test --release --features "private-net runtime-benchmarks"
+    else
+        printf "⚡️ Running Tests for code coverage only\n"
+        export RUSTFLAGS="-Cinstrument-coverage"
+        export SKIP_WASM_BUILD=1
+        export LLVM_PROFILE_FILE="sora2-%p-%m.profraw"
+        rm -rf ~/.cargo/.package-cache
+        cargo fmt -- --check > /dev/null
+        cargo test --features $allfeatures
+    fi
 }
 
 build() {
@@ -44,26 +64,3 @@ build() {
         exit 1
     fi
 }
-
-# build
-# If TAG_NAME is defined, build for a specific tag
-if [[ $buildTag != null ]] && [[ ${TAG_NAME} != null || ${TAG_NAME} != '' ]]; then
-    printf "⚡️ Tag is %s\n" $buildTag ${TAG_NAME}
-    printf "⚡️ Testing with features: private-net runtime-benchmarks\n"
-    releasetest
-    if [[ ${TAG_NAME} =~ 'benchmarking'* ]]; then
-        build 'private-net runtime-benchmarks' 0
-    elif [[ ${TAG_NAME} =~ 'stage'* ]]; then
-        build 'private-net include-real-files ready-to-test' 0
-    elif [[ ${TAG_NAME} =~ 'test'* ]]; then
-        build 'private-net include-real-files reduced-pswap-reward-periods ready-to-test' 0
-    elif [[ -n ${TAG_NAME} && ${TAG_NAME} != 'predev' ]]; then
-        build 'include-real-files' 101
-    fi
-else
-    if [ $prBranch = 'master' ]; then
-        printf "⚡️ Running tests and migrations %s\n"
-        RUST_LOG="debug cargo test --features try-runtime -- run_migrations"
-    fi
-    test
-fi
